refactored code for readability
