+++ subgraphs/euler-finance/src/mappings/handlers.ts
@@ -1,4 +1,4 @@
-import { Address, BigDecimal, BigInt, log } from "@graphprotocol/graph-ts";
+import { Address, BigInt } from "@graphprotocol/graph-ts";
 import {
   AssetStatus,
   Borrow,
@@ -12,12 +12,10 @@ import {
 } from "../../generated/euler/Euler";
 import {
   getCurrentEpoch,
-  getCutoffValue,
   getDeltaStakeAmount,
   getOrCreateAssetStatus,
   getOrCreateLendingProtocol,
   getOrCreateMarket,
-  getOrCreateRewardToken,
   getOrCreateToken,
   getStartBlockForEpoch,
 } from "../common/getters";
@@ -30,15 +28,6 @@ import {
   BIGDECIMAL_ZERO,
   DEFAULT_DECIMALS,
   BIGINT_SEVENTY_FIVE,
-  START_EPOCH,
-  EUL_DIST,
-  EUL_DECIMALS,
-  RewardTokenType,
-  EUL_ADDRESS,
-  BLOCKS_PER_DAY,
-  BLOCKS_PER_EPOCH,
-  BIGINT_ONE,
-  EUL_MARKET_ADDRESS,
 } from "../common/constants";
 import { snapshotFinancials, snapshotMarket, updateUsageMetrics, updateWeightedStakedAmount } from "./helpers";
 import {
@@ -54,7 +43,7 @@ import {
 import { LendingProtocol, Market, Token } from "../../generated/schema";
 import { ERC20 } from "../../generated/euler/ERC20";
 import { GovConvertReserves, GovSetReserveFee } from "../../generated/euler/Exec";
-import { BigDecimalTruncateToBigInt, bigIntChangeDecimals, bigIntToBDUseDecimals } from "../common/conversions";
+import { bigIntChangeDecimals, bigIntToBDUseDecimals } from "../common/conversions";
 import { _Epoch } from "../../generated/schema";
 import { Stake } from "../../generated/EulStakes/EulStakes";
 
@@ -331,102 +320,9 @@ export function handleStake(event: Stake): void {
   let epoch = _Epoch.load(epochID.toString());
   if (!epoch) {
     //Start of a new epoch
-    epoch = new _Epoch(epochID.toString());
+    const epoch = new _Epoch(epochID.toString());
     epoch.epoch = epochID;
     epoch.save();
-
-    // rank markets in the epoch just ended (prev epoch)
-    // find the top ten staked markets; according to the euler guage
-    // https://app.euler.finance/gaugeweight
-    const prevEpochID = epochID - 1;
-    const prevEpoch = _Epoch.load(prevEpochID.toString());
-    if (prevEpoch) {
-      const protocol = getOrCreateLendingProtocol();
-      // finalize mkt._weightedStakedAmount for prev epoch & distribute rewards
-      // The array is needed to select top 10 staked markets
-      const marketWeightedStakedAmounts: BigInt[] = [];
-      for (let i = 0; i < protocol._marketIDs!.length; i++) {
-        const mktID = protocol._marketIDs![i];
-        const mkt = Market.load(mktID);
-        if (!mkt) {
-          log.error("[handleStake]market {} doesn't exist, but this should not happen at tx ={}", [
-            mktID,
-            event.transaction.hash.toHexString(),
-          ]);
-          continue;
-        }
-        const stakedAmount = mkt._stakedAmount;
-        if (stakedAmount.gt(BIGINT_ZERO)) {
-          // finalized mkt._weightedStakedAmount for the epoch just ended
-          // epochStartBlock.minus(BIGINT_ONE) is the end block of prev epoch
-          updateWeightedStakedAmount(mkt, epochStartBlock.minus(BIGINT_ONE));
-        }
-        marketWeightedStakedAmounts.push(mkt._weightedStakedAmount ? mkt._weightedStakedAmount! : BIGINT_ZERO);
-        mkt.save();
-      }
-
-      // update EUL token price
-      const eulerContract = Euler.bind(Address.fromString(EULER_ADDRESS));
-      const execProxyAddress = eulerContract.moduleIdToProxy(MODULEID__EXEC);
-      const eulMarket = getOrCreateMarket(EUL_MARKET_ADDRESS);
-      let EULPriceUSD = updatePrices(execProxyAddress, eulMarket, event);
-      if (!EULPriceUSD) {
-        const EULToken = getOrCreateToken(Address.fromString(EUL_ADDRESS));
-        EULPriceUSD = EULToken.lastPriceUSD!;
-      }
-
-      const rewardToken = getOrCreateRewardToken(Address.fromString(EUL_ADDRESS), RewardTokenType.BORROW);
-      const totalRewardAmount = BigDecimal.fromString((EUL_DIST[prevEpochID - START_EPOCH] * EUL_DECIMALS).toString());
-      // select top 10 staked markets, calculate sqrt(weighted staked amount)
-      const cutoffAmount = getCutoffValue(marketWeightedStakedAmounts, 10);
-      let sumAccumulator = BIGDECIMAL_ZERO;
-      for (let i = 0; i < marketWeightedStakedAmounts.length; i++) {
-        // TODO: reflect changes in eIP 24 and 28
-        if (marketWeightedStakedAmounts[i].ge(cutoffAmount)) {
-          sumAccumulator = sumAccumulator.plus(marketWeightedStakedAmounts[i].sqrt().toBigDecimal());
-        }
-      }
-
-      // scale to daily emission amount
-      const dailyScaler = BigDecimal.fromString((BLOCKS_PER_DAY / (BLOCKS_PER_EPOCH as f64)).toString());
-      for (let i = 0; i < protocol._marketIDs!.length; i++) {
-        const mktID = protocol._marketIDs![i];
-        const mkt = Market.load(mktID);
-        if (!mkt) {
-          log.error("[handleStake]market {} doesn't exist, but this should not happen | tx ={}", [
-            mktID,
-            event.transaction.hash.toHexString(),
-          ]);
-          continue;
-        }
-        if (mkt.rewardTokens && mkt.rewardTokens!.length > 0) {
-          // reset reward emissions for the epoch
-          mkt.rewardTokenEmissionsAmount = [BIGINT_ZERO];
-          mkt.rewardTokenEmissionsUSD = [BIGDECIMAL_ZERO];
-        }
-
-        // distribute the rewards among top 10 staked markets
-        // Only for epochs after START_EPOCH (6)
-        const _weightedStakedAmount = mkt._weightedStakedAmount;
-        if (_weightedStakedAmount && _weightedStakedAmount.ge(cutoffAmount)) {
-          mkt.rewardTokens = [rewardToken.id];
-          const rewardTokenEmissionsAmount = BigDecimalTruncateToBigInt(
-            _weightedStakedAmount.sqrt().divDecimal(sumAccumulator).times(totalRewardAmount).times(dailyScaler),
-          );
-          const rewardTokenEmissionsUSD = rewardTokenEmissionsAmount
-            .divDecimal(BigDecimal.fromString(EUL_DECIMALS.toString()))
-            .times(EULPriceUSD);
-          mkt.rewardTokenEmissionsAmount = [rewardTokenEmissionsAmount];
-          mkt.rewardTokenEmissionsUSD = [rewardTokenEmissionsUSD];
-        }
-
-        // reset mkt._weightedStakedAmount for the new epoch
-        mkt._weightedStakedAmount = BIGINT_ZERO;
-        // EUL staked remains staked for the market until unstaked
-        // so not reset mkt._stakedAmount
-        mkt.save();
-      }
-    }
   }
 
   // In a valid epoch (6 <= epoch <=96) with uninitialized market._stakeLastUpdateBlock
